review fixes №1

+++ 01_week/tasks/addition/addition.cpp
@@ -1,7 +1,5 @@
 #include <cstdint>
-#include <stdexcept>
-
 
 int64_t Addition(int a, int b) {
-    return static_cast<int64_t>(a) + static_cast<int64_t>(b);
+    return static_cast<int64_t>(a) + b;
 }
